detect and handle when containers exit without the Docker Engine stopping or killing them (#36)
Previously only container "start" and "stop" events were handled,
but if a container exited due to an error or if killed by a signal,
whalewall wouldn't detect that and delete appropriate firewall rules.
Listening for the "die" event instead of the "stop" event should
allow whalewall to handle a container stopping for any reason, as
Docker sends a container "die" event no matter how a container was
stopped.