recreate missing nftables rules for containers on startup (#86)
When whalewall first starts, checks are made to ensure that containers
that are no longer running have their rules removed, and containers
that were started while whalewall wasn't running have their rules
created. If whalewall previously created rules for a container,
whalewall was stopped and the container's rules were removed, whalewall
would see that the running container is in the database and do nothing
once whalewall was started again.

Now whalewall will add any missing rules for known containers on
startup. In the future this may be done periodically.